Removed warnings for emx, and legacy makefile fixes

+++ fexist.c
@@ -198,7 +198,7 @@ int _fast direxist(char *directory)
 
     /* Root directory of any drive always exists! */
 
-    if ((isalpha(tempstr[0]) && tempstr[1] == ':' && (tempstr[2] == '\\' || tempstr[2] == '/') &&
+    if ((isalpha((int)tempstr[0]) && tempstr[1] == ':' && (tempstr[2] == '\\' || tempstr[2] == '/') &&
       !tempstr[3]) || eqstr(tempstr, "\\"))
     {
         free(tempstr);
